# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/points": {
            "get": {
                "tags": [
                    "Point"
                ],
                "summary": "getPointList",
                "description": "Get all Points",
                "operationId": "414e326e5df569e950277de2c8e9c3b1",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Point"
                ],
                "summary": "createPoint",
                "description": "Create Point",
                "operationId": "7b9b6b981e45df08007c3f1a42a57595",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/points/{id}": {
            "get": {
                "tags": [
                    "Point"
                ],
                "summary": "getPointItem",
                "description": "Get Point",
                "operationId": "84890aa276d26faec598361de21ed829",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Point",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Point"
                ],
                "summary": "updatePoint",
                "description": "Update Point",
                "operationId": "5a5298550ce46a9468da1fc44a5851e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Point",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Point"
                ],
                "summary": "deletePoint",
                "description": "Delete Point",
                "operationId": "8382eb01398d1029539a633a18910721",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Point",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    }
}